exception handling

+++ server_files.py
@@ -10,50 +10,57 @@ from logger import logger
 worlds_path = path.join(getcwd(), worlds_path)
 template_path = path.join(getcwd(), template_path)
 
-class ServerFolderError(Exception):
-    pass
-
 def copy_template(directory):
     try:
+        if path.exists(directory):
+            rmtree(directory)
         copytree(template_path, directory)
     except OSError:
-        logger.erro('Failed to copy the server template.')
-        raise ServerFolderError('Failed to copy the server template.')
+        logger.warn('Failed to copy the server template.')
 
 def delete_template_copy(directory):
     try:
-        rmtree(directory)
+        if path.exists(directory):
+            rmtree(directory)
     except OSError:
-        logger.erro('Failed to delete a server.')
-        raise ServerFolderError('Failed to delete a server.')
+        logger.warn('Failed to delete a server.')
 
 def delete_whitelist(directory):
     try:
-        remove(path.join(directory, 'whitelist.json'))
+        d = path.join(directory, 'whitelist.json')
+        if path.isfile(d):
+            remove(d)
     except OSError:
         logger.warn('Failed to delete "whitelist.json".')
 
 def delete_ops(directory):
     try:
-        print(path.join(directory, 'ops.json'))
-        remove(path.join(directory, 'ops.json'))
+        d = path.join(directory, 'ops.json')
+        if path.isfile(d):
+            remove(d)
     except OSError:
         logger.warn('Failed to delete "ops.json".')
 
 def delete_world(directory):
     try:
-        rmtree(path.join(directory, 'world'))
+        d = path.join(directory, 'world')
+        if path.exists(d):
+            rmtree(d)
     except OSError:
         logger.warn('Failed to delete the world folder.')
 
-def store_world(directory):
+def store_world(directory, seed):
     if not path.exists(worlds_path):
         try:
             makedirs(worlds_path)
         except OSError:
             pass
-
-    copytree(path.join(directory, 'world'), path.join(worlds_path, f'world-{datetime.now().strftime("%Y-%m-%d-%H-%M-%S")}'))
+    
+    d = path.join(worlds_path, seed)
+    if path.exists(d):
+        rmtree(d)
+        logger.warn('World with the same seed was overwritten.')
+    copytree(path.join(directory, 'world'), d)
